gitea_repos_master/isphere-services/monitoring-api/requirements.txt
python-dotenv==0.20.0
uvicorn==0.15.0
fastapi==0.68.1
keydb==0.0.1
pymongo==4.2.0
async-timeout==4.0.2